staus & genres sync entities methods

+++ services/content-service/src/main/java/dev/animedia/contentservice/application/genre/dto/GenreDto.java
@@ -5,6 +5,6 @@ import java.util.Set;
 public record GenreDto(
 	Long id,
 	String alias,
-	long sortOrder,
+	int sortOrder,
 	Set<GenreTranslationDto> translationSet
 ) {}
